Increase player tracking range
There were some issues where after changing player's skin they would
have duplicates of themselves. One reason that may happen is because
when we change skins, we send the player spawn packet regardless of
whether the player is actually in the range. As it is quite difficult to
determine whether player is in range of another, we can instead ensure
that all players are in each others range at all times.