Do not send skin change related spawn packets to player not in range
This fixes the bug where players saw ghost players - players who did not
actually exist but just stayed at one place. One hypothesis why that
happens was that the skin changed sent player spawn packets to all
players, even those that were not in range of the player.

Now, we check if the entity tracker for the player whose skin is being
changed tracks the player that the packet is sent to. If not, then the
spawn packet is not sent.