Specify map and randomization factor in sumo demo script
We recently started requiring the user to also specify the map on which
to play on. This commit fixes the demo script which broke because it did
not specify these parameters.

Randomization factor is always set to 1. Map can be configured by user.
If the user does not configure it, it is picked randomly.