Do not load models from all iterations into memory
Previously, we loaded all models to memory because the opponent was
picked from one of these models. After a lot of iterations, the models
start to take considerable amount of memory (around 20 GB) which
sometimes caused training to crash.

To fix the issue with memory, we keep 50 models loaded in memory at all
times that can be given to environments. This approach solves the
problem with memory usage but at the same time keeps the time it takes
to retrieve a model low.