Keep chunks loaded at all times
In classic duels, when a chunk is not loaded when the game starts, the
player can get stuck inside a block and die. To prevent that, the
following is done:
- Chunks are now never unloaded. That is, even if no player is close to
a chunk, it is still kept loaded.
- The classic duel chunks where players are teleported to are loaded
when the server starts. This is done to avoid the game immediately
ending when a map is played for the first time.