Change skins in classic after short period
There seems to be some kind of bug where players do not see each other
and I suspect it is related to the skin changing logic. Since in sumo
skins are not that important (because players are wearing armor), we can
delay changing the skins 100 ticks into the game. Further, players are
quite far from each other in the start anyway.